requirements.txt
flask
pymondo
dnspython
python_dotenv